Changed long long int to int64_t for explicitness

+++ src/lib/Money.cpp
@@ -3,18 +3,18 @@
 
 namespace fmc {
 
-    Money::Money(long long int dollars_, int cents_) : dollars{dollars_}, cents{cents_} {
+    Money::Money(int64_t dollars_, int cents_) : dollars{dollars_}, cents{cents_} {
         this->normalize();
     }
     Money::Money(double v) {
-        long long int total = std::llround(v * 100);
+        int64_t total = std::llround(v * 100);
         this->dollars = total / 100;
         this->cents = (int) total % 100;
         this->normalize();
     }
 
     void Money::normalize() {
-        long long int total = this->dollars * 100 + this->cents;
+        int64_t total = this->dollars * 100 + this->cents;
 
         this->dollars = total / 100;
         this->cents = (int) total % 100;
@@ -59,7 +59,7 @@ namespace fmc {
     }
 
     Money Money::operator*(uint i) const {
-        long long int total = this->dollars * 100 + this->cents;
+        int64_t total = this->dollars * 100 + this->cents;
         total *= i;
 
         Money result{total / 100, (int) total % 100};
@@ -67,8 +67,8 @@ namespace fmc {
 
         return result;
     }
-    Money Money::operator*(long long int i) const {
-        long long int total = this->dollars * 100 + this->cents;
+    Money Money::operator*(int64_t i) const {
+        int64_t total = this->dollars * 100 + this->cents;
         total *= i;
 
         Money result{total / 100, (int) total % 100};
@@ -87,8 +87,8 @@ namespace fmc {
     }
 
     double Money::operator/(const Money& m) const {
-        long long int cur_total = this->dollars * 100 + this->cents;
-        long long int other_total = m.dollars * 100 + m.cents;
+        int64_t cur_total = this->dollars * 100 + this->cents;
+        int64_t other_total = m.dollars * 100 + m.cents;
 
         return ((double) cur_total) / ((double) other_total);
     }
